django_admin/requirements.txt
django==4.1.3
django-cleanup==6.0.0
django-cors-headers==3.13.0
django-debug-toolbar==3.7.0
django-extensions==3.2.1
django-postgres-extra==2.0.6
django-split-settings==1.2.0
flake8==5.0.4
psycopg2==2.9.5
python-dotenv==0.21.0
uwsgi==2.0.21
